Improve diagnosis for non-existing components
When a non-themable component is listed in the include list, such as `FONT_ICON`, or on a typo, the `path.join` that produces `componentPath` produces an error that gives no clue as to the actual problem.  This fix detects the error, reports which is the wrong/misspelled component and lists the available ones.